+++ example_project/lib/example_web/components/layouts/app.html.heex
@@ -45,40 +45,46 @@
         5
       </a>
       <a
-        href={~p"/plus-minus-svelte"}
+        href={~p"/sigil"}
         class="font-semibold leading-6 text-zinc-900 hover:text-zinc-700 hover:underline p-2"
       >
         6
       </a>
       <a
-        href={~p"/plus-minus-liveview"}
+        href={~p"/plus-minus-svelte"}
         class="font-semibold leading-6 text-zinc-900 hover:text-zinc-700 hover:underline p-2"
       >
         7
       </a>
       <a
-        href={~p"/plus-minus-hybrid"}
+        href={~p"/plus-minus-liveview"}
         class="font-semibold leading-6 text-zinc-900 hover:text-zinc-700 hover:underline p-2"
       >
         8
       </a>
       <a
-        href={~p"/log-list"}
+        href={~p"/plus-minus-hybrid"}
         class="font-semibold leading-6 text-zinc-900 hover:text-zinc-700 hover:underline p-2"
       >
         9
       </a>
       <a
-        href={~p"/breaking-news"}
+        href={~p"/log-list"}
         class="font-semibold leading-6 text-zinc-900 hover:text-zinc-700 hover:underline p-2"
       >
         10
       </a>
       <a
-        href={~p"/chat"}
+        href={~p"/breaking-news"}
         class="font-semibold leading-6 text-zinc-900 hover:text-zinc-700 hover:underline p-2"
       >
         11
+      </a>
+      <a
+        href={~p"/chat"}
+        class="font-semibold leading-6 text-zinc-900 hover:text-zinc-700 hover:underline p-2"
+      >
+        12
       </a>
     </div>
     <div class="flex items-center gap-4">
